requirements.txt
openai>=1.0.0
PyYAML>=6.0
rich>=13.0.0
reportlab>=4.0.0 